Revert "rename to pycapture"
This reverts commit 37a3255658e69c1d6af7b1ad0cc21e42b6ddc339.
