+++ README.md
@@ -16,7 +16,7 @@ Features:
 ### Install
 You can install this tool via pip:
 ```
-pip install pycapture
+pip install pyhttpcap
 ```
 
 ### Analysis pcap file
