+++ pyhttpcap/constant.py
