+++ pyhttpcap/reader.py
