+++ pyhttpcap/textutils.py
